revert test update

+++ python/test/unit/runtime/test_blaslt.py
@@ -22,7 +22,7 @@ def test_blaslt(m, n, k, dtype_str, device):
             pytest.skip("float8_e4m3fnuz is only supported on HIP CDNA3")
         if dtype_str == "float8_e4m3fn" and not is_hip_cdna4():
             pytest.skip("float8_e4m3fn is only supported on HIP CDNA4")
-        c_dtype = torch.bfloat16 if dtype_str in ("float8_e4m3fnuz", "float8_e4m3fn") else dtype
+        c_dtype = torch.float16 if dtype_str in ("float8_e4m3fnuz", "float8_e4m3fn") else dtype
         make_handle = lambda workspace: vendor.hipblas.HipblasLt(workspace)
     else:
         pytest.skip("test_blaslt is only supported on CUDA or HIP")
